+++ PSAzureMigrationAdvisor/internal/scripts/variables.ps1
@@ -1,0 +1,2 @@
+﻿# Mapping data used to add additional metadata to the script report export
+$script:migrationCommandMapping = @{ }